# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University API",
        "description": "UniversityAPI Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/students": {
            "get": {
                "tags": [
                    "students"
                ],
                "summary": "Get list of students",
                "description": "Returns list of students",
                "operationId": "getStudentssList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "students"
                ],
                "summary": "Create student",
                "description": "Create the student in store",
                "operationId": "addStudent",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fac_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "students"
                ],
                "summary": "Get student by Id",
                "description": "Returns the student by Id",
                "operationId": "getStudentbyId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Student not found"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "students"
                ],
                "summary": "Update student by Id",
                "description": "Update the student in store",
                "operationId": "updateStudent",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful updated"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "students"
                ],
                "summary": "Delete student by Id",
                "description": "Delete the student from store",
                "operationId": "deleteStudent",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    }
}